Split out sendTransactionBatch
